requirements.txt
paho-mqtt
requests
schedule
kafka-python
pymongo
